generic WES run working - needs more UI work
